+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Yii2 multiple input change log
 1.2.20 in development
 ---------------------
 
+- #79 Added support for embedded MultipleInput widget (unclead, execut)
+
 1.2.19
 ------